use hpp for c++ code
for consistency and to not confuse cppcheck
