+++ rmw_fastrtps_cpp/include/rmw_fastrtps_cpp/MessageTypeSupport.hpp
@@ -12,44 +12,31 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef RMW_FASTRTPS_CPP__SERVICETYPESUPPORT_H_
-#define RMW_FASTRTPS_CPP__SERVICETYPESUPPORT_H_
+#ifndef RMW_FASTRTPS_CPP__MESSAGETYPESUPPORT_H_
+#define RMW_FASTRTPS_CPP__MESSAGETYPESUPPORT_H_
 
 #include <fastcdr/FastBuffer.h>
 #include <fastcdr/Cdr.h>
+
 #include <cassert>
+#include <memory>
 
-#include "TypeSupport.h"
+#include "TypeSupport.hpp"
+#include "rosidl_typesupport_introspection_cpp/message_introspection.hpp"
 #include "rosidl_typesupport_introspection_cpp/field_types.hpp"
-
-struct CustomServiceInfo;
 
 namespace rmw_fastrtps_cpp
 {
 
 template<typename MembersType>
-class ServiceTypeSupport : public TypeSupport<MembersType>
-{
-protected:
-  ServiceTypeSupport();
-};
-
-template<typename ServiceMembersType, typename MessageMembersType>
-class RequestTypeSupport : public ServiceTypeSupport<MessageMembersType>
+class MessageTypeSupport : public TypeSupport<MembersType>
 {
 public:
-  explicit RequestTypeSupport(const ServiceMembersType * members);
-};
-
-template<typename ServiceMembersType, typename MessageMembersType>
-class ResponseTypeSupport : public ServiceTypeSupport<MessageMembersType>
-{
-public:
-  explicit ResponseTypeSupport(const ServiceMembersType * members);
+  explicit MessageTypeSupport(const MembersType * members);
 };
 
 }  // namespace rmw_fastrtps_cpp
 
-#include "ServiceTypeSupport_impl.h"
+#include "MessageTypeSupport_impl.hpp"
 
-#endif  // RMW_FASTRTPS_CPP__SERVICETYPESUPPORT_H_
+#endif  // RMW_FASTRTPS_CPP__MESSAGETYPESUPPORT_H_
